refactor: Move authInfo interface to Authorization folder
